refactor(app): split modal workflow reducers
