+++ src/app/update/modal/mod.rs
@@ -1,514 +1,39 @@
+mod base;
+mod confirm_dialog;
+mod er_picker;
+mod help;
+mod query_history;
+mod settings;
+
 use std::time::Instant;
 
-use crate::catalog::HelpDocument;
-use crate::cmd::effect::Effect;
 use crate::model::app_state::AppState;
-use crate::model::shared::confirm_dialog::ConfirmIntent;
-use crate::model::shared::flash_timer::FlashId;
-use crate::model::shared::help::HelpOrigin;
-use crate::model::shared::input_mode::InputMode;
-use crate::ports::outbound::AppSettings;
-use crate::update::action::{
-    Action, InputTarget, ListMotion, ListTarget, ModalKind, ScrollAmount, ScrollDirection,
-    ScrollTarget,
-};
+use crate::update::action::Action;
 use crate::update::dispatch_result::DispatchResult;
 
-fn scroll_help_by(
-    state: &mut AppState,
-    direction: ScrollDirection,
-    delta: usize,
-    line_count: usize,
-    content_width: usize,
-) {
-    let max_scroll = state.ui.help_max_scroll(line_count, content_width);
-    let offset = direction.clamp_vertical_offset(state.ui.help.scroll_offset(), max_scroll, delta);
-    state.ui.help.set_scroll_offset(offset);
+#[cfg(test)]
+use crate::cmd::effect::Effect;
+#[cfg(test)]
+use crate::model::shared::input_mode::InputMode;
+#[cfg(test)]
+use crate::ports::outbound::AppSettings;
+#[cfg(test)]
+use crate::update::action::{
+    InputTarget, ListMotion, ListTarget, ModalKind, ScrollAmount, ScrollDirection, ScrollTarget,
+};
+
+pub fn dispatch_modal(state: &mut AppState, action: &Action, now: Instant) -> DispatchResult {
+    base::reduce_base_lifecycle(state, action, now)
+        .or_else(|| settings::reduce_settings(state, action, now))
+        .or_else(|| help::reduce_help(state, action, now))
+        .or_else(|| confirm_dialog::reduce_confirm_dialog(state, action, now))
+        .or_else(|| er_picker::reduce_er_picker(state, action, now))
+        .or_else(|| query_history::reduce_query_history_picker(state, action, now))
 }
 
-fn scroll_help_horizontally(state: &mut AppState, direction: ScrollDirection) {
-    match direction {
-        ScrollDirection::Left => {
-            state
-                .ui
-                .help
-                .set_horizontal_offset(state.ui.help.horizontal_offset().saturating_sub(1));
-        }
-        ScrollDirection::Right => {
-            let document = HelpDocument::from_state(state);
-            let max_scroll = state
-                .ui
-                .help_max_horizontal_scroll(document.line_count(), document.content_width());
-            state
-                .ui
-                .help
-                .set_horizontal_offset((state.ui.help.horizontal_offset() + 1).min(max_scroll));
-        }
-        ScrollDirection::Up | ScrollDirection::Down => {}
-    }
-}
-
-fn close_help(state: &mut AppState) {
-    state.modal.pop_mode();
-    state.ui.help.close();
-}
-
-pub fn reduce_modal(state: &mut AppState, action: &Action, now: Instant) -> DispatchResult {
-    match action {
-        Action::OpenModal(ModalKind::TablePicker) => {
-            state.modal.set_mode(InputMode::TablePicker);
-            state.ui.table_picker.clear_filter_and_reset();
-            DispatchResult::handled()
-        }
-        Action::CloseModal(ModalKind::TablePicker)
-        | Action::CloseModal(ModalKind::CommandPalette)
-        | Action::Escape => {
-            state.modal.set_mode(InputMode::Normal);
-            DispatchResult::handled()
-        }
-        Action::OpenModal(ModalKind::CommandPalette) => {
-            state.modal.set_mode(InputMode::CommandPalette);
-            // Command palette currently reuses the generic picker selection state.
-            state.ui.table_picker.reset();
-            DispatchResult::handled()
-        }
-        Action::OpenModal(ModalKind::Settings) => {
-            state.settings.open(state.ui.theme_id());
-            state.modal.set_mode(InputMode::Settings);
-            DispatchResult::handled()
-        }
-        Action::SettingsSelectNext => {
-            state.settings.select_next();
-            DispatchResult::handled()
-        }
-        Action::SettingsSelectPrevious => {
-            state.settings.select_previous();
-            DispatchResult::handled()
-        }
-        Action::SettingsNextSection => {
-            state.settings.switch_next_section();
-            DispatchResult::handled()
-        }
-        Action::SettingsPreviousSection => {
-            state.settings.switch_previous_section();
-            DispatchResult::handled()
-        }
-        Action::SettingsStartCustomBrowserEdit => {
-            state.settings.start_custom_browser_edit();
-            DispatchResult::handled()
-        }
-        Action::SettingsStopCustomBrowserEdit => {
-            state.settings.stop_custom_browser_edit();
-            DispatchResult::handled()
-        }
-        Action::TextInput {
-            target: InputTarget::SettingsErBrowser,
-            ch,
-        } => {
-            state.settings.input_custom_browser(*ch);
-            DispatchResult::handled()
-        }
-        Action::TextBackspace {
-            target: InputTarget::SettingsErBrowser,
-        } => {
-            state.settings.backspace_custom_browser();
-            DispatchResult::handled()
-        }
-        Action::TextDelete {
-            target: InputTarget::SettingsErBrowser,
-        } => {
-            state.settings.delete_custom_browser();
-            DispatchResult::handled()
-        }
-        Action::TextMoveCursor {
-            target: InputTarget::SettingsErBrowser,
-            direction,
-        } => {
-            state.settings.move_custom_browser_cursor(*direction);
-            DispatchResult::handled()
-        }
-        Action::SettingsApply => {
-            let theme_id = state.settings.selected_theme();
-            let settings = AppSettings {
-                theme_id,
-                er_browser: state.settings.selected_er_browser(),
-            };
-            DispatchResult::handled_with(vec![Effect::SaveSettings { settings }])
-        }
-        Action::SettingsCancel | Action::CloseModal(ModalKind::Settings) => {
-            state.settings.discard_selection();
-            state.modal.set_mode(InputMode::Normal);
-            DispatchResult::handled()
-        }
-        Action::SettingsSaved(settings) => {
-            state.ui.set_theme(settings.theme_id);
-            state
-                .settings
-                .commit_saved(settings.theme_id, settings.er_browser.clone());
-            state.set_success("Settings saved".to_string());
-            DispatchResult::handled()
-        }
-        Action::SettingsSaveFailed(error) => {
-            state.set_error(format!("Failed to save settings: {error}"));
-            DispatchResult::handled()
-        }
-        Action::ToggleModal(ModalKind::Help) => {
-            if state.modal.active_mode() == InputMode::Help {
-                close_help(state);
-            } else {
-                let origin = HelpOrigin::from_state(state);
-                state.ui.help.open(origin);
-                state.modal.push_mode(InputMode::Help);
-            }
-            DispatchResult::handled()
-        }
-        Action::CloseModal(ModalKind::Help) => {
-            close_help(state);
-            DispatchResult::handled()
-        }
-        Action::TextInput {
-            target: InputTarget::HelpFilter,
-            ch,
-        } => {
-            state.ui.help.insert_filter_char(*ch);
-            DispatchResult::handled()
-        }
-        Action::TextBackspace {
-            target: InputTarget::HelpFilter,
-        } => {
-            state.ui.help.backspace_filter();
-            DispatchResult::handled()
-        }
-        Action::Scroll {
-            target: ScrollTarget::Help,
-            direction,
-            amount,
-        } => {
-            match amount {
-                ScrollAmount::Line
-                    if matches!(direction, ScrollDirection::Left | ScrollDirection::Right) =>
-                {
-                    scroll_help_horizontally(state, *direction);
-                }
-                ScrollAmount::Line => {
-                    let document = HelpDocument::from_state(state);
-                    scroll_help_by(
-                        state,
-                        *direction,
-                        1,
-                        document.line_count(),
-                        document.content_width(),
-                    );
-                }
-                ScrollAmount::ToStart => {
-                    if matches!(direction, ScrollDirection::Left | ScrollDirection::Right) {
-                        state.ui.help.set_horizontal_offset(0);
-                    } else {
-                        state.ui.help.set_scroll_offset(0);
-                    }
-                }
-                ScrollAmount::ToEnd => {
-                    let document = HelpDocument::from_state(state);
-                    if matches!(direction, ScrollDirection::Left | ScrollDirection::Right) {
-                        let max_scroll = state.ui.help_max_horizontal_scroll(
-                            document.line_count(),
-                            document.content_width(),
-                        );
-                        state.ui.help.set_horizontal_offset(max_scroll);
-                    } else {
-                        let max_scroll = state
-                            .ui
-                            .help_max_scroll(document.line_count(), document.content_width());
-                        state.ui.help.set_scroll_offset(max_scroll);
-                    }
-                }
-                ScrollAmount::HalfPage | ScrollAmount::FullPage => {
-                    let document = HelpDocument::from_state(state);
-                    let line_count = document.line_count();
-                    let content_width = document.content_width();
-                    if let Some(delta) =
-                        amount.page_delta(state.ui.help_visible_rows(line_count, content_width))
-                    {
-                        scroll_help_by(state, *direction, delta, line_count, content_width);
-                    }
-                }
-                ScrollAmount::ViewportTop
-                | ScrollAmount::ViewportMiddle
-                | ScrollAmount::ViewportBottom => {}
-            }
-            DispatchResult::handled()
-        }
-        Action::Scroll {
-            target: ScrollTarget::ConfirmDialog,
-            direction,
-            amount: ScrollAmount::Line,
-        } => {
-            let max_scroll = state.confirm_dialog.max_scroll() as usize;
-            state.confirm_dialog.preview_scroll = direction.clamp_vertical_offset(
-                state.confirm_dialog.preview_scroll as usize,
-                max_scroll,
-                1,
-            ) as u16;
-            DispatchResult::handled()
-        }
-        Action::CloseModal(ModalKind::SqlModal) => {
-            state.modal.set_mode(InputMode::Normal);
-            state.sql_modal.cleanup_on_close();
-            state.flash_timers.clear(FlashId::SqlModal);
-            DispatchResult::handled()
-        }
-        Action::OpenModal(ModalKind::ErTablePicker) => {
-            if state.session.metadata().is_none() {
-                state.ui.pending_er_picker = true;
-                state.set_success("Waiting for metadata...".to_string());
-                return DispatchResult::handled();
-            }
-            state.ui.pending_er_picker = false;
-            state.ui.er_selected_tables.clear();
-            state.modal.set_mode(InputMode::ErTablePicker);
-            state.ui.er_picker.clear_filter_and_reset();
-            DispatchResult::handled()
-        }
-        Action::CloseModal(ModalKind::ErTablePicker) => {
-            state.modal.set_mode(InputMode::Normal);
-            state.ui.er_picker.clear_filter();
-            state.ui.er_selected_tables.clear();
-            state.ui.pending_er_picker = false;
-            DispatchResult::handled()
-        }
-        Action::TextInput {
-            target: InputTarget::ErFilter,
-            ch: c,
-        } => {
-            state.ui.er_picker.insert_filter_char(*c);
-            DispatchResult::handled()
-        }
-        Action::TextBackspace {
-            target: InputTarget::ErFilter,
-        } => {
-            state.ui.er_picker.backspace_filter();
-            DispatchResult::handled()
-        }
-        Action::TextMoveCursor {
-            target: InputTarget::ErFilter,
-            direction: movement,
-        } => {
-            state.ui.er_picker.move_filter_cursor(*movement);
-            DispatchResult::handled()
-        }
-        Action::ErToggleSelection => {
-            let filtered = state.er_filtered_tables();
-            if let Some(table) = filtered.get(state.ui.er_picker.selected()) {
-                let name = table.qualified_name();
-                if !state.ui.er_selected_tables.remove(&name) {
-                    state.ui.er_selected_tables.insert(name);
-                }
-            }
-            DispatchResult::handled()
-        }
-        Action::ErSelectAll => {
-            let all_tables: Vec<String> =
-                state.tables().iter().map(|t| t.qualified_name()).collect();
-            if state.ui.er_selected_tables.len() == all_tables.len() {
-                state.ui.er_selected_tables.clear();
-            } else {
-                state.ui.er_selected_tables = all_tables.into_iter().collect();
-            }
-            DispatchResult::handled()
-        }
-        Action::ErConfirmSelection => {
-            if state.ui.er_selected_tables.is_empty() {
-                state.set_error("No tables selected".to_string());
-                return DispatchResult::handled();
-            }
-            state.er_preparation.target_tables =
-                state.ui.er_selected_tables.iter().cloned().collect();
-            state.modal.set_mode(InputMode::Normal);
-            state.ui.er_picker.clear_filter();
-            state.ui.er_selected_tables.clear();
-            DispatchResult::handled_with(vec![Effect::DispatchActions(vec![Action::ErOpenDiagram])])
-        }
-        Action::OpenModal(ModalKind::QueryHistoryPicker) => {
-            if state.session.active_connection_id.is_none() {
-                return DispatchResult::handled();
-            }
-            if state.query.is_running() {
-                return DispatchResult::handled();
-            }
-            if state.modal.active_mode() == InputMode::ConfirmDialog {
-                return DispatchResult::handled();
-            }
-            if state.sql_modal.completion().visible
-                && !state.sql_modal.completion().candidates.is_empty()
-            {
-                return DispatchResult::handled();
-            }
-
-            state.query_history_picker.reset();
-            state.modal.push_mode(InputMode::QueryHistoryPicker);
-
-            let conn_id = state.session.active_connection_id.as_ref().unwrap();
-            DispatchResult::handled_with(vec![Effect::LoadQueryHistory {
-                project_name: state.runtime.project_name.clone(),
-                connection_id: conn_id.clone(),
-            }])
-        }
-        Action::CloseModal(ModalKind::QueryHistoryPicker) => {
-            state.modal.pop_mode();
-            state.query_history_picker.reset();
-            DispatchResult::handled()
-        }
-        Action::QueryHistoryLoaded(conn_id, entries) => {
-            if state.modal.active_mode() != InputMode::QueryHistoryPicker {
-                return DispatchResult::handled();
-            }
-            if state.session.active_connection_id.as_ref() != Some(conn_id) {
-                return DispatchResult::handled();
-            }
-            state.query_history_picker.replace_entries(entries);
-            DispatchResult::handled()
-        }
-        Action::QueryHistoryLoadFailed(e) => {
-            if state.modal.active_mode() != InputMode::QueryHistoryPicker {
-                return DispatchResult::handled();
-            }
-            state.messages.set_error_at(e.to_string(), now);
-            DispatchResult::handled()
-        }
-        Action::QueryHistoryAppendFailed(_) => DispatchResult::handled(),
-        Action::TextInput {
-            target: InputTarget::QueryHistoryFilter,
-            ch: c,
-        } => {
-            state.query_history_picker.insert_filter_char(*c);
-            DispatchResult::handled()
-        }
-        Action::TextBackspace {
-            target: InputTarget::QueryHistoryFilter,
-        } => {
-            state.query_history_picker.backspace_filter();
-            DispatchResult::handled()
-        }
-        Action::ListSelect {
-            target: ListTarget::QueryHistory,
-            motion: ListMotion::Next,
-        } => {
-            state.query_history_picker.select_next();
-            DispatchResult::handled()
-        }
-        Action::ListSelect {
-            target: ListTarget::QueryHistory,
-            motion: ListMotion::Previous,
-        } => {
-            state.query_history_picker.select_previous();
-            DispatchResult::handled()
-        }
-        Action::QueryHistoryConfirmSelection => {
-            let grouped = state.query_history_picker.grouped_filtered_entries();
-            let selected = state.query_history_picker.clamped_selected();
-            let query = grouped.get(selected).map(|g| g.entry.query.clone());
-            let origin = state.modal.pop_mode();
-
-            state.query_history_picker.reset();
-
-            let Some(query) = query else {
-                return DispatchResult::handled();
-            };
-
-            match origin {
-                InputMode::Normal => {
-                    state.modal.set_mode(InputMode::SqlModal);
-                    state.sql_modal.load_query_from_history(query);
-                }
-                InputMode::SqlModal => {
-                    state.sql_modal.load_query_from_history(query);
-                }
-                _ => {}
-            }
-            DispatchResult::handled()
-        }
-
-        Action::ConfirmDialogConfirm => {
-            let intent = state.confirm_dialog.take_intent();
-            state.modal.pop_mode();
-
-            match intent {
-                Some(ConfirmIntent::QuitNoConnection) => {
-                    state.should_quit = true;
-                    DispatchResult::handled()
-                }
-                Some(ConfirmIntent::DeleteConnection(id)) => {
-                    DispatchResult::handled_with(vec![Effect::DeleteConnection { id }])
-                }
-                Some(ConfirmIntent::ExecuteWrite { blocked: true, .. }) => {
-                    state.result_interaction.clear_write_preview();
-                    state.query.clear_delete_refresh_target();
-                    DispatchResult::handled()
-                }
-                Some(ConfirmIntent::ExecuteWrite {
-                    sql,
-                    blocked: false,
-                }) => {
-                    if let Some(dsn) = &state.session.dsn {
-                        state.query.begin_running(now);
-                        DispatchResult::handled_with(vec![Effect::ExecuteWrite {
-                            dsn: dsn.clone(),
-                            query: sql,
-                            read_only: state.session.read_only,
-                        }])
-                    } else {
-                        state.result_interaction.clear_write_preview();
-                        state.query.clear_delete_refresh_target();
-                        state
-                            .messages
-                            .set_error_at("No active connection".to_string(), now);
-                        DispatchResult::handled()
-                    }
-                }
-                Some(ConfirmIntent::DisableReadOnly) => {
-                    state.session.read_only = false;
-                    DispatchResult::handled()
-                }
-                Some(ConfirmIntent::CsvExport {
-                    export_query,
-                    file_name,
-                    row_count,
-                }) => {
-                    if let Some(dsn) = &state.session.dsn {
-                        DispatchResult::handled_with(vec![Effect::ExportCsv {
-                            dsn: dsn.clone(),
-                            query: export_query,
-                            file_name,
-                            row_count,
-                            read_only: state.session.read_only,
-                        }])
-                    } else {
-                        DispatchResult::handled()
-                    }
-                }
-                None => DispatchResult::handled(),
-            }
-        }
-        Action::ConfirmDialogCancel => {
-            let intent = state.confirm_dialog.take_intent();
-            state.result_interaction.clear_write_preview();
-            state.query.clear_delete_refresh_target();
-
-            if matches!(intent, Some(ConfirmIntent::QuitNoConnection)) {
-                state.connection_setup.reset();
-                if !state.connections().is_empty() || state.session.dsn.is_some() {
-                    state.connection_setup.is_first_run = false;
-                }
-                state.modal.pop_mode_override(InputMode::ConnectionSetup);
-                DispatchResult::handled()
-            } else {
-                state.modal.pop_mode();
-                DispatchResult::handled()
-            }
-        }
-
-        _ => DispatchResult::pass(),
-    }
+#[cfg(test)]
+fn reduce_modal(state: &mut AppState, action: &Action, now: Instant) -> DispatchResult {
+    dispatch_modal(state, action, now)
 }
 
 #[cfg(test)]
